+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Jackson 2.9.7
     * AWS 1.11.414
     * Apache CXF 3.2.6
+* BREAKING: Change `AReferenceableEntity` to interface. Extend `AEntity` additionally.
 
 # Version 1.28
 * Update dependencies